refactor: renamed feature folders as use-case names
- MakeAnOfferForm renamed to MakeOffer
- ListingsPage renamed to ViewListings
