--- src/domains/tasks/MakeAnOfferForm/components/index.tsx
@@ -1,2 +1,0 @@
-export * from "./MakeAnOfferForm";
-export * from "./MakeAnOfferModal";
