--- src/domains/tasks/MakeAnOfferForm/hooks/index.tsx
@@ -1,1 +1,0 @@
-export * from "./useMakeAnOfferForm";
