+++ src/domains/tasks/MakeOffer/components/MakeOfferModal.tsx
@@ -1,5 +1,5 @@
 import { Modal } from "@shopify/polaris";
-import { MakeAnOfferForm } from "./MakeAnOfferForm";
+import { MakeAnOfferForm } from "./MakeOfferForm";
 import { createOffer } from "../api";
 import { MakeAnOfferFormState } from "../hooks";
