+++ src/domains/tasks/MakeOffer/components/index.tsx
@@ -1,0 +1,2 @@
+export * from "./MakeOfferForm";
+export * from "./MakeOfferModal";
