+++ src/domains/tasks/components/TaskCategoryBadge.tsx
@@ -1,7 +1,7 @@
 import { Badge } from "@shopify/polaris";
 import { Status } from "@shopify/polaris/build/ts/latest/src/components/Badge";
 
-import { TaskStatus, CategoryStatusMap } from "@Tasks/ListingsPage/types";
+import { TaskStatus, CategoryStatusMap } from "@Tasks/ViewListings/types";
 
 export type TaskCategoryBadgeProps = {
   status: TaskStatus;
